Improved model performance: 0.913 F1-macro (was 0.787)
- Retrained XGB+LGBM ensemble with regularization (max_depth=4, reg_alpha/lambda)
- Class-balanced sample weights (3x real upweight) + synthetic 1x synthetic)
- Removed leakage feature 'malicious_family' (was 45% gain importance)
- Added 'has_malicious_vt_tags' to synthetic data & features
- Learned ensemble weights (XGB=0.64, LGB=0.36) via differential evolution
- Temperature calibration (T=1.1) for confidence reliability
- CRITICAL confidence floor (0.75) prevents false CRITICAL alerts
- Low-confidence ML fallback to risk model fixes LOW-on-malicious bug
- Groq-powered dynamic recommended actions per investigation
- IOC-level LRU cache eliminates non-deterministic repeated calls
- Auto-demo mode when no API keys configured
- All 30 tests pass

+++ utils/risk_model.py
@@ -15,7 +15,6 @@ import re
 from dataclasses import dataclass
 from pathlib import Path
 
-
 MODEL_PATH = Path(__file__).resolve().parents[1] / "models" / "risk_model.json"
 
 FEATURES = [
@@ -213,7 +212,7 @@ def predict_risk(text: str, model: dict | None = None, ml_features: dict | None 
 
 
 def _compute_confidence(risk_score: float, ml_features: dict | None = None) -> int:
-    base = int(round(55 + abs(risk_score - 0.5) * 90))
+    base = round(55 + abs(risk_score - 0.5) * 90)
     if ml_features is None:
         return base
     data_coverage = sum([
